requirements.txt
# Core LLM and embedding functionality
openai>=1.14.3
langchain>=0.1.15
langchain-core>=0.1.50
langchain-community>=0.0.32
langchain-openai>=0.1.6

# Streamlit for UI
streamlit>=1.34.0

# FAISS for vector store (CPU version)
faiss-cpu>=1.7.4

# URL loader dependency
unstructured[all-docs]>=0.10.2

# Optional: dotenv for loading environment variables
python-dotenv>=1.0.0

# Compatibility fix: Pydantic v1 for older LangChain versions (optional, if needed)
# pydantic<2.0
